64bit zobrist hash

+++ src/native/common.h
@@ -3,8 +3,8 @@ typedef unsigned short uint16_t;
 typedef unsigned int   uint32_t;
 #pragma once
 
-typedef unsigned long  uint64_t;
-typedef unsigned long  size_t;
+typedef unsigned long long uint64_t;
+typedef unsigned long size_t;
 typedef short int16_t;
 typedef int   int32_t;
 typedef int bool;
